@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { BehaviorSubject, combineLatest, merge, Observable } from 'rxjs';
 import { filter, map, startWith } from 'rxjs/operators';
@@ -26,7 +26,7 @@ export class ParamsComponent extends AbstractSubscribeUnsubscribeDirective imple
   trackById = (i: number, playground: Playground): string => playground.id;
 
   constructor(
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private router: Router,
     private route: ActivatedRoute,
     private service: PlaygroundService,
